Store OrientedCoordinateArrays in unordered_map
This improves performance by reducing the number of
OrientedCoordinateArray comparisons needed with std::set.

+++ include/geos/geom/Envelope.h
@@ -453,7 +453,7 @@ public:
      */
     double distance(const Envelope* env) const;
 
-    int hashCode() const;
+    size_t hashCode() const;
 
 private:
